pyproject.toml
[project]
name = "eeg"
version = "0.1.0"
description = "EEG preprocessing and exercises"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "fonttools==4.60.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "kiwisolver==1.4.9",
    "lazy-loader==0.4",
    "markupsafe==3.0.3",
    "matplotlib==3.10.7",
    "mne==1.10.2",
    "mne-bids==0.17.0",
    "numpy==2.3.4",
    "packaging==25.0",
    "pillow==12.0.0",
    "platformdirs==4.5.0",
    "pooch==1.8.2",
    "pyyaml==6.0.3",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "requests==2.32.5",
    "scipy==1.16.2",
    "six==1.17.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "urllib3==2.5.0",
    "pandas==2.3.3",
    "pyqt6>=6.10.1",
    "mne-qt-browser>=0.7.4",
    "psutil>=7.1.3",
    "scikit-learn==1.8.0",
    "torch>=2.9.1",
    "joblib==1.5.3",
    "pyside6==6.10.1",
    "pyside6-addons==6.10.1",
    "pyside6-essentials==6.10.1",
    "pytz==2025.2",
    "shiboken6==6.10.1",
    "threadpoolctl==3.6.0",
    "tzdata==2025.2",
    "seaborn>=0.13.2",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.uv]
package = false
